Preoptimization, Errors, Visualization, Quality

+++ backend/experiments.py
@@ -29,8 +29,8 @@ from werkzeug.utils import secure_filename
 #sys.path.insert(1, '/Classes/')
 from backend import MLA
 from backend import Validation
-from backend import Preoptimization
-from backend.calculator import Calculator
+from backend.Classes.PreoptimizationFiles import Preoptimization
+from backend.Classes.PreoptimizationFiles import Standardization
 
 from sklearn.model_selection import KFold
 
@@ -45,13 +45,16 @@ logger = logging.getLogger()
 def load_possible_experiments():
     list_Algorithms = MLA.getMLAs()
     list_Preoptimizations = Preoptimization.getPreopts()
+    list_Preopt_Categories = Preoptimization.getPreopt_Categories()
+    list_Standardization = Standardization.getStandardization()
 
     #----------------new Cycon--------------------------->
     # First choice is between various categories of ML 
-    g.section_Method = ["MLA: Machine Learning Algorithm", "DLANN: Deep Learning Artifical Neural Networks"]
+    g.section_Method = ["MLA", "DLANN"]
+    g.section_Method_Display_Name = ["Machine Learning Algorithm (MLA)", "Deep Learning Artifical Neural Networks (DLANN)"]
     g.section_Info = [["Machine learning algorithms are mathematical model mapping methods. They are used to learn patterns embedded in the existing training dataset in order to perform pattern recognition, classification, and prediction.\n\nCurrently, the only algorithms available on cycon is under the classification objective. Other objectives that will be added later include clustering and regression."],
                       ["DLANN info"]]
-    g.Methodologies = zip(g.section_Method, g.section_Info)
+    g.Methodologies = zip(g.section_Method, g.section_Method_Display_Name, g.section_Info)
 
     # Obtain the selection of Algorithm Names and Definition.
     algorithm_Names = []
@@ -68,17 +71,53 @@ def load_possible_experiments():
 
     # obtain the selection of preoptimization options and definitions.
     preopt_Names = []
+    preopt_Display_Names = []
     preopt_Definition = []
+
     preopt_Names.append("")
+    preopt_Display_Names.append("")
     preopt_Definition.append("")
 
     for preopt in list_Preoptimizations:
         preopt_Names.append(preopt.getName())
+        preopt_Display_Names.append(preopt.getDisplayName())
         preopt_Definition.append(preopt.getDefinition())
     g.section_preopt = preopt_Names
+    g.section_preopt_names = preopt_Display_Names
     g.section_Info = preopt_Definition
-    g.Preoptimizations = zip(g.section_preopt, g.section_Info)
+    g.Preoptimizations = zip(g.section_preopt, g.section_preopt_names, g.section_Info)
 
+    # create list of options to choose a category of preoptimization options.
+    preopt_Category_Names = []
+    preopt_Category_Display_Names = []
+    preopt_Category_Definition = []
+
+    # preopt_Category_Names.append("")
+    # preopt_Category_Display_Names.append("")
+    # preopt_Category_Definition.append("")
+
+    for preopt_Category in list_Preopt_Categories:
+        preopt_Category_Names.append(preopt_Category.getName())
+        preopt_Category_Display_Names.append(preopt_Category.getDisplayName())
+        preopt_Category_Definition.append(preopt_Category.getDefinition())
+    g.section_preopt_Category_Names = preopt_Category_Names
+    g.section_preopt_Category_Display_Names = preopt_Category_Display_Names
+    g.section_preopt_Category_Info =  preopt_Category_Definition
+    g.preopt_Categories = zip(g.section_preopt_Category_Names, g.section_preopt_Category_Display_Names, g.section_preopt_Category_Info)
+
+    # create list of options for the first category of preoprtimization (I.E. Standardization)
+    preopt_Names = []
+    preopt_Display_Names = []
+    preopt_Definition = []
+
+    for preopt in list_Standardization:
+        preopt_Names.append(preopt.getName())
+        preopt_Display_Names.append(preopt.getDisplayName())
+        preopt_Definition.append(preopt.getDefinition())
+    g.section_preopt_Names = preopt_Names
+    g.section_preopt_Display_Names = preopt_Display_Names
+    g.section_preopt_Info =  preopt_Definition
+    g.preopts = zip(g.section_preopt_Names, g.section_preopt_Display_Names, g.section_preopt_Info)
 
 
 
@@ -87,17 +126,9 @@ def load_possible_experiments():
 def LCA_Old():
     return render_template("experiments/LCA_Old.html")
 
-
-# REMOVE ASAP
-@bp.route("/new_experiment")
-def new_experiment():
-    return render_template("experiments/new_experiment.html")
-
-
 @bp.route("/lca")
 def lca():
     return render_template("experiments/lca.html")
-
 
 @bp.route("/cycon")
 def cycon():
@@ -130,34 +161,35 @@ def run_experiment():
 
     # Split
     if data['validation'] == "Split":
-        Metrics = Validation.Split(data)
+        status, msg, Metrics = Validation.Split(data)
 
     # K-Fold
     elif data['validation'] == "K-Fold":
-        Metrics = Validation.K_Fold(data)
+        status, msg, Metrics = Validation.K_Fold(data)
 
+    if status == "worked":
+        # Open json file for the experiment.
+        baseFolder = os.getcwd()
+        locationSavedResults = Path(baseFolder) / "SavedResults"
+        filename = secure_filename(data["projectName"] + ".json")
+        filepath = locationSavedResults / filename.lower()
+        if os.path.exists(filepath):
+            os.remove(filepath)
+            fp = open(filepath, "a")
+        else:
+            fp = open(filepath, "a")
     
-    # Open json file for the experiment.
-    baseFolder = os.getcwd()
-    locationSavedResults = Path(baseFolder) / "SavedResults"
-    filename = secure_filename(data["projectName"] + ".json")
-    filepath = locationSavedResults / filename.lower()
-    if os.path.exists(filepath):
-        os.remove(filepath)
-        fp = open(filepath, "a")
-    else:
-        fp = open(filepath, "a")
-    
-    # write to json file
-    metrics_Dump = json.dumps(Metrics)
+        # write to json file
+        metrics_Dump = json.dumps(Metrics)
 
-    fp.write(metrics_Dump)
+        fp.write(metrics_Dump)
 
-    # close the connection
-    fp.close()
+        # close the connection
+        fp.close()
 
-    return json.dumps(Metrics)
+    Results = [status, msg, Metrics]
 
+    return json.dumps(Results)
 
 @bp.route("/getResults", methods=["POST"])
 def getResults(): 
@@ -171,7 +203,7 @@ def getResults():
     filename = data["projectName"] + ".json"
     filepath = locationSavedResults / filename.lower()
     fp = open(filepath, "r")
-    
+
     Metrics = json.load(fp)
 
     # close the connection
@@ -189,6 +221,15 @@ def getAlgorithmParameters():
 
     return json.dumps(Parameters)
 
+@bp.route("/getCategoryPreopts", methods=["POST"])
+def getCategoryPreopts():
+    output = request.get_json()
+    data = json.loads(output)
+
+    preopts = Preoptimization.getCategoryPreopts(data["Category"])
+
+    return json.dumps(preopts)
+
 
 @bp.route("/results")
 def results():
@@ -202,6 +243,27 @@ def getCSVResults():
 
     data = formData['form']
 
-    pdfs = Preoptimization.getCSV_PDF(data)
+    status, msg, info = Preoptimization.getCSV_PDF(data)
 
-    return json.dumps(pdfs)
+    Results = [status, msg, info]
+
+    return json.dumps(Results)
+
+@bp.route("/getPreoptParameters", methods=["POST"])
+def getPreoptParameters():
+    output = request.get_json()
+    data = json.loads(output)
+
+    Parameters = Preoptimization.getParameters(data["Preopt"])
+
+    return json.dumps(Parameters)
+
+# Gets the colummn names inside the csv.
+@bp.route("/getCSVColumnTitles", methods=["POST"])
+def getCSVColumnTitles():
+    output = request.get_json()
+    data = json.loads(output)
+
+    columnTitles = Preoptimization.getCSVColumnTitles(data["csvFileName"])
+
+    return json.dumps(columnTitles)
